+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Fix XML templates overriding and reloading
 - `title` and `description` are now optional in the front matter
 - Add GenericConfig, Vim syntax
+- Add `_index.md` for homepage as well
+- Allow sorting by `none`, `date` and `order` for sections
 
 ## 0.0.4 (2017-04-23)